Replace if...else statement inside getPlayerChoice with a while loop
Previous implementation inside getPlayerChoice was based on using an
undeclared variable in order for the function to always return the
user input. Using a global variable inside a function is considered
bad practice.
The function now executes a while loop using a declared variable and
loops the user through prompt until a condition is met.