Create convertPlayerChoice function
convertPlayerChoice initializes the prompt and converts the input to
uppercase only if the cancel button is not clicked. Converting the
prompt value to uppercase directly in the checkPlayerChoice function
threw an error when the value was null (Cancel == null) as toUpperCase()
is not being able to read an object.
Now convertPlayerChoice is making sure that the value that is provided
to checkPlayerChoice for further checks is either an uppercase string,
an empty string or a null value.