Introduce score functionality
The variables are declared globally and are being incremented in
playRound function based on the if...else if conditions.
An if statement
inside the startGame() loop displays the score as long as the prompt
input does not return null, else it breaks the loop and ends the loop
setting the score of both players to null.
The if...else...else statement outside the loop of startGame() outputs
the winner based on the score - if the score is null for both players,
the else statement string from playRound() is displayed, as the game
was cancelled by the player.
Finally, the score variables are declared inside startGame() after being
declared globally therefore when the function is called to start a new
game the score is being reset.